added metric for how much we have to fix the local env
ideally we want this to be 0

+++ link_bot_classifiers/src/link_bot_classifiers/nn_classifier2.py
@@ -25,6 +25,7 @@ from moonshine.classifier_losses_and_metrics import class_weighted_mean_loss
 from moonshine.metrics import BinaryAccuracyOnPositives, BinaryAccuracyOnNegatives, LossMetric, \
     FalsePositiveMistakeRate, FalseNegativeMistakeRate, FalsePositiveOverallRate, FalseNegativeOverallRate, BinaryRate
 from moonshine.moonshine_utils import numpify
+from tensorflow.keras.metrics import Mean
 from moonshine.my_keras_model import MyKerasModel
 from visualization_msgs.msg import MarkerArray, Marker
 
@@ -109,7 +110,7 @@ class NNClassifier(MyKerasModel):
         batch_size = inputs['batch_size']
         time = inputs['time']
 
-        if debug_input():
+        if debug_input() and training:
             # clear the other voxel grids from previous calls
             self.debug.clear()
             self.scenario.delete_points_rviz(label='attract')
@@ -144,7 +145,7 @@ class NNClassifier(MyKerasModel):
         voxel_grids = self.vg_info.make_voxelgrid_inputs(inputs, local_env, local_origin_point, batch_size, time)
         inputs['voxel_grids'] = voxel_grids
 
-        if debug_input():
+        if debug_input() and training:
             assert not rospy.get_param("use_sim_time", False)
             self.debug_viz_inputs(inputs, local_origin_point, time)
 
@@ -216,6 +217,7 @@ class NNClassifier(MyKerasModel):
             'fp/total':              FalsePositiveOverallRate(),
             'fn/total':              FalseNegativeOverallRate(),
             'aug_validity_rate':     BinaryRate(),
+            'local_env_aug_fix_delta': Mean(),
         }
 
     def compute_metrics(self, metrics: Dict[str, Metric], losses: Dict, dataset_element, outputs):
@@ -232,6 +234,7 @@ class NNClassifier(MyKerasModel):
         metrics['accuracy on positives'].update_state(y_true=labels, y_pred=probabilities)
         if self.aug.do_augmentation() and self.aug.is_valids is not None:
             metrics['aug_validity_rate'].update_state(self.aug.is_valids)
+            metrics['local_env_aug_fix_delta'].update_state(self.aug.local_env_aug_fix_delta)
 
     def conv_encoder(self, voxel_grids, batch_size, time):
         conv_outputs_array = tf.TensorArray(tf.float32, size=0, dynamic_size=True)
